fix button types circ dep
